Polish header styling
Changed nav element to a div, because the buttons don't serve
navigational purposes.

Improved button sizing to be more responsive. Previously, button text
was getting crushed because the button was being sized with a percentage
value. Switched to rem units for consistency across screen heights.

Made the .search-container div a flex box in order to properly center
the search bar. Made the search bar slightly larger.