Add overflow rules to main element
The page was overflowing, causing scroll bars to appear on each axis.
This was not ideal. The main-container's height was set to 100vh, and
the main element's height was set to fill 100% of that height. So now,
whenever content in the main element overflows beyond the vertical
viewport, it creates a vertical scroll bar within the main element
instead of on the webpage as a whole.